test: split broad regression modules
Replace the remaining multi-topic regression files with
focused test modules for flakeref resolution, nix argv
construction, target resolution, CLI error boundaries,
temp SBOM generation, and SBOM vulnerability enrichment.

At the same time, move broad tests off common.utils and
onto the focused common modules directly. Keep common.utils
only as a small external compatibility shim, with one
dedicated compatibility test file and an explicit module
note.
