+++ tests/test_flakeref_resolution.py
@@ -1,0 +1,146 @@
+#!/usr/bin/env python3
+# SPDX-FileCopyrightText: 2026 Technology Innovation Institute (TII)
+#
+# SPDX-License-Identifier: Apache-2.0
+# pylint: disable=missing-function-docstring
+
+"""Focused tests for flakeref resolution helpers."""
+
+from types import SimpleNamespace
+
+import pytest
+
+from common.errors import FlakeRefRealisationError, FlakeRefResolutionError
+from common.flakeref import try_resolve_flakeref
+
+
+def test_try_resolve_flakeref_uses_argv_lists():
+    calls = []
+
+    def fake_exec_cmd(cmd, **kwargs):
+        calls.append((cmd, kwargs))
+        if cmd[1] == "eval":
+            return SimpleNamespace(stdout="/nix/store/resolved\n", returncode=0)
+        return SimpleNamespace(stdout="", stderr="", returncode=0)
+
+    resolved = try_resolve_flakeref(
+        "/tmp/my flake#pkg",
+        force_realise=True,
+        impure=True,
+        exec_cmd_fn=fake_exec_cmd,
+    )
+
+    assert resolved == "/nix/store/resolved"
+    assert calls == [
+        (
+            [
+                "nix",
+                "eval",
+                "--raw",
+                "/tmp/my flake#pkg",
+                "--extra-experimental-features",
+                "flakes",
+                "--extra-experimental-features",
+                "nix-command",
+                "--impure",
+            ],
+            {"raise_on_error": False, "return_error": True, "log_error": False},
+        ),
+        (
+            [
+                "nix",
+                "build",
+                "--no-link",
+                "/tmp/my flake#pkg",
+                "--extra-experimental-features",
+                "flakes",
+                "--extra-experimental-features",
+                "nix-command",
+                "--impure",
+            ],
+            {"raise_on_error": False, "return_error": True, "log_error": False},
+        ),
+    ]
+
+
+def test_try_resolve_flakeref_raises_on_failed_force_realise():
+    def fake_exec_cmd(cmd, **_kwargs):
+        if cmd[1] == "eval":
+            return SimpleNamespace(stdout="/nix/store/resolved\n", returncode=0)
+        return SimpleNamespace(stdout="", stderr="build failed", returncode=1)
+
+    with pytest.raises(FlakeRefRealisationError, match="build failed"):
+        try_resolve_flakeref(
+            "/tmp/my flake#pkg",
+            force_realise=True,
+            exec_cmd_fn=fake_exec_cmd,
+        )
+
+
+def test_try_resolve_flakeref_raises_on_failed_eval_for_flakeref():
+    def fake_exec_cmd(_cmd, **_kwargs):
+        return SimpleNamespace(stdout="", stderr="attribute missing", returncode=1)
+
+    with pytest.raises(FlakeRefResolutionError, match="attribute missing"):
+        try_resolve_flakeref(".#missing", exec_cmd_fn=fake_exec_cmd)
+
+
+def test_try_resolve_flakeref_returns_none_for_non_flake_path():
+    def fake_exec_cmd(_cmd, **_kwargs):
+        return SimpleNamespace(
+            stdout="",
+            stderr="does not contain a 'flake.nix'",
+            returncode=1,
+        )
+
+    resolved = try_resolve_flakeref(
+        "/nix/store/not-a-flake-output",
+        exec_cmd_fn=fake_exec_cmd,
+    )
+
+    assert resolved is None
+
+
+@pytest.mark.parametrize("path", ["missing", "./missing", "foo/bar"])
+def test_try_resolve_flakeref_returns_none_for_missing_relative_paths(path):
+    def fake_exec_cmd(_cmd, **_kwargs):
+        return SimpleNamespace(stdout="", stderr="dummy eval failure", returncode=1)
+
+    resolved = try_resolve_flakeref(path, exec_cmd_fn=fake_exec_cmd)
+
+    assert resolved is None
+
+
+def test_try_resolve_flakeref_returns_none_for_existing_fragment_path_when_eval_fails(
+    tmp_path,
+):
+    existing_path = tmp_path / "contains#hash"
+    existing_path.mkdir()
+    calls = []
+
+    def fake_exec_cmd(cmd, **kwargs):
+        calls.append((cmd, kwargs))
+        return SimpleNamespace(
+            stdout="",
+            stderr="does not contain a 'flake.nix'",
+            returncode=1,
+        )
+
+    resolved = try_resolve_flakeref(existing_path.as_posix(), exec_cmd_fn=fake_exec_cmd)
+
+    assert resolved is None
+    assert calls
+
+
+def test_flakeref_realisation_error_accepts_none_stderr():
+    error = FlakeRefRealisationError(".#pkg", stderr=None)
+
+    assert error.stderr == ""
+    assert str(error) == "Failed force-realising flakeref '.#pkg'"
+
+
+def test_flake_ref_resolution_error_preserves_stderr_verbatim():
+    error = FlakeRefResolutionError(".#pkg", stderr="stderr details\n")
+
+    assert error.stderr == "stderr details\n"
+    assert str(error) == "Failed evaluating flakeref '.#pkg': stderr details"
